feat: apply formatting and rename vue components
