+++ gigadb/app/client/web/src/components/DatasetInfo.vue
@@ -1,0 +1,16 @@
+<template>
+  <h4>
+    GigaDB: {{ stage }} for the dataset {{ identifier }}
+  </h4>
+</template>
+<style></style>
+<script>
+export default {
+  props: ['identifier'],
+  data: function () {
+    return {
+      stage: "Uploading files",
+    }
+  },
+}
+</script>
