+++ gigadb/app/client/web/src/plugins/uppy-checksum.js
@@ -4,7 +4,7 @@ const ChunkedFileReader = require('chunked-file-reader')
 const SparkMD5 = require('spark-md5')
 
 export const Checksum = class Checksum extends BasePlugin {
-  constructor (uppy, opts) {
+  constructor(uppy, opts) {
     super(uppy, opts)
     this.id = opts.id || 'Checksum'
     this.type = 'modifier'
@@ -13,11 +13,11 @@ export const Checksum = class Checksum extends BasePlugin {
     this.calcMd5 = this.calcMd5.bind(this)
   }
 
-  calcMd5 (file) {
+  calcMd5(file) {
     this.uppy.log(`[Checksum] Calculating Hash: ${file.name}`)
     return new Promise(function (resolve, reject) {
-      var spark  = new SparkMD5.ArrayBuffer(),
-          reader = new ChunkedFileReader();
+      var spark = new SparkMD5.ArrayBuffer(),
+        reader = new ChunkedFileReader();
 
       reader.subscribe('chunk', function (e) {
         spark.append(e.chunk);
@@ -25,7 +25,7 @@ export const Checksum = class Checksum extends BasePlugin {
 
       reader.subscribe('end', function (e) {
         // var rawHash    = spark.end(true);
-        var rawHash    = spark.end();
+        var rawHash = spark.end();
         // var base64Hash = btoa(rawHash);
         // this.uppy.log(rawHash)
         // resolve(base64Hash);
@@ -36,7 +36,7 @@ export const Checksum = class Checksum extends BasePlugin {
     })
   }
 
-  prepareUpload (fileIDs) {
+  prepareUpload(fileIDs) {
     const promises = fileIDs.map((fileID) => {
       const file = this.uppy.getFile(fileID)
       this.uppy.log(`[Checksum] prepareUpload for file ${file.name}`)
@@ -51,11 +51,11 @@ export const Checksum = class Checksum extends BasePlugin {
     return Promise.all(promises)
   }
 
-  install () {
+  install() {
     this.uppy.addPreProcessor(this.prepareUpload)
   }
 
-  uninstall () {
+  uninstall() {
     this.uppy.removePreProcessor(this.prepareUpload)
   }
 }
